dev_requirements.txt
-r requirements.txt
ipython[notebook]
django-extensions
django-test-without-migrations
